+++ CHANGELOG.rst
@@ -24,6 +24,7 @@ Release 0.11.0 (unreleased)
 * Add more tests and documentation for patching (#888)
 * Restrict ``src`` to string only in schema (#888)
 * Don't consider ignored files for determining local changes (#350)
+* Add filter command (#19)
 
 Release 0.10.0 (released 2025-03-12)
 ====================================